Тысячи копипасты с бея и тг. Звук пробивания стены, улучшен счётчик фото в камере, лярвы правильно растут от мышей, красивое исчезание дыма, новые спрайты, новые причёски (проверка), изменение доступа к айди консоли, stars() proc теперь использует html_decode(), починен метаболизм с капсаицином, убраны объявления о конверте майнда в борге (looks like they just reverted to their old faith), изменены задачи трейторов, карта проверена через debug и подкорректирована, возможность переворачивать столы, в т.ч. покерный, усилены лазеры и эмитеры

+++ code/game/objects/effects/effect_system.dm
@@ -36,6 +36,17 @@ would spawn and follow the beaker, even if it is carried or thrown.
 		reagents.delete()
 	return
 
+/datum/effect/effect/proc/fadeOut(var/atom/A, var/frames = 16) //This is a piece of TG code, takes advantage of 500+ versions use of alpha-particles. This will make smoke gradually disappear.
+	if(A.alpha == 0) //Handle already transparent case
+		return
+	if(frames == 0)
+		frames = 1 //We will just assume that by 0 frames, the coder meant "during one frame".
+	var/step = A.alpha / frames
+	for(var/i = 0, i < frames, i++)
+		A.alpha -= step
+		sleep(world.tick_lag)
+	return
+
 
 /obj/effect/effect/water/New()
 	..()
@@ -281,7 +292,9 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(75+rand(10,30))
-					if (smoke) smoke.delete()
+					if (smoke)
+						smoke.delete()
+						fadeOut(smoke)
 					src.total_smoke--
 
 
@@ -381,7 +394,9 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if (smoke)
+						smoke.delete()
+						fadeOut(smoke)
 					src.total_smoke--
 
 
@@ -510,7 +525,9 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if (smoke)
+						smoke.delete()
+						fadeOut(smoke)
 					src.total_smoke--
 
 
@@ -605,7 +622,9 @@ steam.start() -- spawns the effect
 					sleep(10)
 					step(smoke,direction)
 				spawn(150+rand(10,30))
-					smoke.delete()
+					if (smoke)
+						smoke.delete()
+						fadeOut(smoke)
 					src.total_smoke--
 
 /////////////////////////////////////////////
